map(K_CITY_20260618): link_61/86/87/88 정수 필드 dtype 수정 (genpy 직렬화)
신규 생성된 4개 링크가 정수 필드를 float64로 저장 → to_control_team 발행 시
struct.error(field is_stop_line must be an integer type)로 실차에서 직렬화 실패.

- mat 4개: 정수 필드 12개 float64→uint8(정상 링크와 동일),
  look_at_IntersectionID는 1500(link_87) 수용 위해 msg 정의대로 uint16.
  east/north/station은 float64 유지. 값 손실 없음, serialize() 재현 테스트 통과.
- convert_links.py: 생성기도 동일 dtype으로 저장하도록 수정(재발 방지).

+++ src/localization/gps_system_localizer/mapfiles/K_CITY_20260618/convert_links.py
@@ -70,18 +70,20 @@ def main():
         mat_data = {
             'east':                     east,
             'north':                    north,
-            'LINK_ID':                  np.array([[cfg['link_id']]], dtype=np.float64),
-            'NEXT_LINK_ID':             np.array([[cfg['next_id']]], dtype=np.float64),
-            'have_to_LangeChange_right': np.array([[0]], dtype=np.float64),
-            'have_to_LangeChange_left':  np.array([[0]], dtype=np.float64),
-            'left_LaneChange_avail':     np.array([[0]], dtype=np.float64),
-            'right_LaneChange_avail':    np.array([[0]], dtype=np.float64),
-            'RIGHT_LINK_ID':             np.array([[0]], dtype=np.float64),
-            'LEFT_LINK_ID':              np.array([[0]], dtype=np.float64),
-            'is_stop_line':              np.array([[cfg['is_stop']]], dtype=np.float64),
-            'Speed_Limit':               np.array([[SPEED_LIMIT]], dtype=np.float64),
-            'look_at_signalGroupID':     np.array([[cfg['sgid']]], dtype=np.float64),
-            'look_at_IntersectionID':    np.array([[cfg['iid']]], dtype=np.float64),
+            # 정수 필드는 uint8 (기존 정상 링크 link_*.mat 와 동일, genpy 직렬화 호환).
+            # look_at_IntersectionID 만 값이 255 초과(예 1500)라 msg 정의대로 uint16.
+            'LINK_ID':                  np.array([[cfg['link_id']]], dtype=np.uint8),
+            'NEXT_LINK_ID':             np.array([[cfg['next_id']]], dtype=np.uint8),
+            'have_to_LangeChange_right': np.array([[0]], dtype=np.uint8),
+            'have_to_LangeChange_left':  np.array([[0]], dtype=np.uint8),
+            'left_LaneChange_avail':     np.array([[0]], dtype=np.uint8),
+            'right_LaneChange_avail':    np.array([[0]], dtype=np.uint8),
+            'RIGHT_LINK_ID':             np.array([[0]], dtype=np.uint8),
+            'LEFT_LINK_ID':              np.array([[0]], dtype=np.uint8),
+            'is_stop_line':              np.array([[cfg['is_stop']]], dtype=np.uint8),
+            'Speed_Limit':               np.array([[SPEED_LIMIT]], dtype=np.uint8),
+            'look_at_signalGroupID':     np.array([[cfg['sgid']]], dtype=np.uint8),
+            'look_at_IntersectionID':    np.array([[cfg['iid']]], dtype=np.uint16),
             'station':                   station,
         }
